+++ packages/app/src/onError.ts
@@ -3,13 +3,10 @@ import { ErrorHandler } from './router'
 import { Request } from './request'
 import { Response } from './response'
 
-export const onErrorHandler: ErrorHandler = (
-  err: any,
-  _req: Request,
-  res: Response
-) => {
-  if (!res.headersSent) {
+export const onErrorHandler: ErrorHandler = (err: any, _req: Request, res: Response) => {
+  if (!res.writableEnded) {
     const code = (res.statusCode = err.code || err.status || 500)
+
     if (typeof err === 'string' || Buffer.isBuffer(err)) res.end(err)
     else res.end(err.message || STATUS_CODES[code])
   }