+++ packages/logger/src/index.ts
@@ -1,54 +1,61 @@
 import colors from 'colors'
-import dayjs from 'dayjs';
+import dayjs from 'dayjs'
 import { IncomingMessage as Request, ServerResponse as Response, METHODS } from 'http'
 
-interface Timestamp {
-  format: string;
+export interface LoggerOptions {
+  methods?: string[]
+  output?: {
+    color: boolean
+    callback: (string) => void
+  }
+  timestamp?: boolean | { format?: string }
 }
 
-interface LoggerProperties {
-  methods?: string[];
-  timestamp?: Timestamp
-}
-
-const loggerHandler = (props: LoggerProperties) => {
-  props = {
-    methods: METHODS,
-    ...props
-  };
+export const logger = (options: LoggerOptions = {}) => {
+  const methods = options.methods ?? METHODS
+  const timestamp = options.timestamp ?? false
+  const output = options.output ?? { callback: console.log, color: true }
 
   const logger = (req: Request, res: Response, next?: () => void) => {
     res.on('finish', () => {
       const { method, url } = req
       const { statusCode, statusMessage } = res
 
-      if (method && props.methods.includes(method)) {
+      if (method && methods.includes(method)) {
         const s = statusCode.toString()
 
         let status: string = s
         let msg: string = statusMessage
 
-        let timestamp = ''
-        if (props.timestamp) {
-          timestamp += `${dayjs().format(props.timestamp.format).toString()} - `
+        let time = ''
+        if (timestamp) {
+          if (typeof timestamp !== 'boolean' && timestamp.format) {
+            time += `${dayjs().format(timestamp.format).toString()} - `
+          } else {
+            time += `${dayjs().format('HH:mm:ss').toString()} - `
+          }
         }
 
-        switch (s[0]) {
-          case '2':
-            status = colors.cyan.bold(s)
-            msg = colors.cyan(msg)
-            console.log(timestamp + `${method} ${status} ${msg} ${url}`)
-            break
-          case '4':
-            status = colors.red.bold(s)
-            msg = colors.red(msg)
-            console.log(timestamp + `${method} ${status} ${msg} ${url}`)
-            break
-          case '5':
-            status = colors.magenta.bold(s)
-            msg = colors.magenta(msg)
-            console.error(timestamp + `${method} ${status} ${msg} ${url}`)
-            break
+        if (!output.color) {
+          output.callback(`${time}${method} ${status} ${msg} ${url}`)
+        } else {
+          switch (s[0]) {
+            case '2':
+              status = colors.cyan.bold(s)
+              msg = colors.cyan(msg)
+              output.callback(`${time}${method} ${status} ${msg} ${url}`)
+              break
+            case '4':
+              status = colors.red.bold(s)
+              msg = colors.red(msg)
+              output.callback(`${time}${method} ${status} ${msg} ${url}`)
+              break
+            case '5':
+              status = colors.magenta.bold(s)
+              msg = colors.magenta(msg)
+              output.callback(`${time}${method} ${status} ${msg} ${url}`)
+              break
+          }
         }
       }
     })
@@ -58,5 +65,3 @@ const loggerHandler = (props: LoggerProperties) => {
 
   return logger
 }
-
-export default loggerHandler