+++ site/server.ts
@@ -1,7 +1,8 @@
 import { App } from '../packages/app/src'
 import serve from 'serve-handler'
 import { markdownStaticHandler as md } from '../packages/markdown/src'
-import logger from '@tinyhttp/logger'
+import { staticHandler } from '../packages/static/src'
+import { logger } from '@tinyhttp/logger'
 import { createReadStream } from 'fs'
 import { transformMWPageStream, transformPageIndexStream } from './streams'
 import unfetch from 'isomorphic-unfetch'
@@ -14,38 +15,6 @@ const NON_MW_PKGS = ['app', 'etag', 'cookie', 'cookie-signature']
 
 app
   .use(logger())
-  .get('/mw', async (req, res, next) => {
-    let json: any, status: number, msg: string
-
-    try {
-      const res = await unfetch('https://api.github.com/repos/talentlessguy/tinyhttp/contents/packages')
-
-      status = res.status
-      msg = res.statusText
-      json = await res.json()
-    } catch (e) {
-      next(e)
-    }
-
-    if (status !== 200) {
-      next(msg)
-    } else {
-      const readStream = createReadStream(`${HTML_PATH}/search.html`)
-
-      let transformer = transformPageIndexStream(json.filter(e => !NON_MW_PKGS.includes(e.name)))
-
-      if (req.query.q) {
-        const results = json.filter((el: any) => {
-          const query = req.query.q as string
-
-          return el.name.indexOf(query.toLowerCase()) > -1
-        })
-        transformer = transformPageIndexStream(results.filter(e => !NON_MW_PKGS.includes(e.name)))
-      }
-
-      readStream.pipe(transformer).pipe(res)
-    }
-  })
   .get('/mw/:mw', async (req, res, next) => {
     if (NON_MW_PKGS.includes(req.params.mw)) {
       next()
@@ -70,20 +39,53 @@ app
       }
     }
   })
+  .get('/mw', async (req, res, next) => {
+    let json: any, status: number, msg: string
+
+    try {
+      const res = await unfetch('https://api.github.com/repos/talentlessguy/tinyhttp/contents/packages')
+
+      status = res.status
+      msg = res.statusText
+      json = await res.json()
+    } catch (e) {
+      next(e)
+    }
+
+    if (status !== 200) {
+      next(msg)
+    } else {
+      const readStream = createReadStream(`${HTML_PATH}/search.html`)
+
+      let transformer = transformPageIndexStream(json.filter((e) => !NON_MW_PKGS.includes(e.name)))
+
+      if (req.query.q) {
+        const results = json.filter((el: any) => {
+          const query = req.query.q as string
+
+          return el.name.indexOf(query.toLowerCase()) > -1
+        })
+        transformer = transformPageIndexStream(results.filter((e) => !NON_MW_PKGS.includes(e.name)))
+      }
+
+      readStream.pipe(transformer).pipe(res)
+    }
+  })
+
   .use(
     md('pages/md', {
       stripExtension: true,
       markedExtensions: [
         {
-          headerIds: true
-        }
-      ]
+          headerIds: true,
+        },
+      ],
     })
   )
-  .use((req, res) =>
-    serve(req, res, {
-      public: 'static'
+  .use(async (req, res) => {
+    await serve(req, res, {
+      public: 'static',
     })
-  )
+  })
 
 app.listen(3000, () => console.log(`Running on http://localhost:3000`))