+++ packages/cookie-parser/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @tinyhttp/cookie-parser
 
+## 0.1.19
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/app@0.2.30
+
+## 0.1.18
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/app@0.2.28
+
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/app@0.2.27
+
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/app@0.2.26
+
+## 0.1.15
+
+### Patch Changes
+
+- Update cookie-parser imports
+
+## 0.1.14
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/app@0.2.21
+
+## 0.1.13
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/app@0.2.19
+
 ## 0.1.12
 
 ### Patch Changes